+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## [Unreleased]
 
-- No changes yet.
+- Fix LSP published diagnostics to filter to the opened file.
+- Add `textDocument/documentSymbol` support for `buf lsp serve`.
 
 ## [v1.59.0] - 2025-10-20